requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
dj-database-url==2.1.0
Django==3.2.1
django-crispy-forms==2.0
django-environ==0.11.2
djangorestframework==3.14.0
gunicorn==21.2.0
importlib-metadata==6.7.0
Jinja2==3.1.3
MarkupSafe==2.1.4
packaging==23.2
psycopg2==2.9.9
Pygments==2.17.2
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
whitenoise==6.5.0
zipp==3.15.0
